1.拍摄流程整理

+++ app/src/main/java/com/ybj366533/yycamera/utils/ToolUtils.kt
@@ -162,7 +162,7 @@ object ToolUtils {
     }
 
     // 格式化时间 HH:MM:SS 或者 MM:SS
-    fun stringForTime(timeMs: Long): String {
+    fun stringForTime(timeMs: Int): String {
         val totalSeconds = timeMs / 1000
 
         val seconds = totalSeconds % 60
